The routes for the admin aswell us the routes for the users are set
