+++ src/main/java/vcet/cse/placement/automation/service/StudentsService.java
@@ -2,27 +2,20 @@ package vcet.cse.placement.automation.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import vcet.cse.placement.automation.Model.Students;
-import vcet.cse.placement.automation.Model.StudentScores;
 import vcet.cse.placement.automation.repo.StudentsDatabaseCollector;
 import vcet.cse.placement.automation.exception.StudentNotFoundException;
 import java.util.*;
 import java.util.stream.Collectors;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
- 
-
- 
 
 @CrossOrigin
 @Service
 public class StudentsService {
 @Autowired
-
-    
-     
-    private StudentsDatabaseCollector studentsDB;
+private StudentsDatabaseCollector studentsDB;
  
-    /* Get all the students data from the data base */
+    /* Get all the students data from the data-base */
 
     public List<Students> getStudents() {
         return studentsDB.findAll();
@@ -71,6 +64,7 @@ public class StudentsService {
         
         studentsDB.saveAll(students);
     }
+
     /* Methos to validate the students data */
 
     private void validateStudent(Students student) {
@@ -103,6 +97,7 @@ public class StudentsService {
  
  
     /* method to update the students data for a particular student */
+
     @Transactional
     public Map<String, Object> updateAllStudentData(List<Students> studentsToUpdate) {
         List<String> errors = new ArrayList<>();
@@ -154,6 +149,8 @@ public class StudentsService {
                 failedUniversityNos.add(studentUpdate.getUniversityNo());
             }
         }
+
+
         
         // Return results
         Map<String, Object> result = new HashMap<>();
@@ -168,6 +165,7 @@ public class StudentsService {
  
 
     /* method to update the student data for a particular student */
+
     @Transactional
     public void updateStudent(Long universityNo, Students student) {
         if (!studentsDB.existsById(universityNo)) {
@@ -197,98 +195,7 @@ public class StudentsService {
     }
 
 
-    /* Method to get the chart data for the home profile page */
-
- 
-    public Map<String, Object> getChartData(int batch) {
-        List<Students> batchStudents = studentsDB.findAll().stream()
-            .filter(student -> student.getBatch() == batch)
-            .collect(Collectors.toList());
-        
-        Map<String, Object> chartData = new HashMap<>();
-        
-        // 1. Donut Chart Data - LeetCode Performance by Class
-        Map<String, Integer> classLeetcodeScores = batchStudents.stream()
-            .collect(Collectors.groupingBy(
-                Students::getClassName,
-                Collectors.summingInt(student -> 
-                    student.getEasyProblemsSolved() + 
-                    student.getMediumProblemsSolved() + 
-                    student.getHardProblemsSolved()
-                )
-            ));
-        
-        // Convert to percentages for the donut chart
-        int totalProblems = classLeetcodeScores.values().stream().mapToInt(Integer::intValue).sum();
-        List<Integer> donutData = new ArrayList<>();
-        if (totalProblems > 0) {
-            donutData = classLeetcodeScores.values().stream()
-                .map(problems -> (problems * 100) / totalProblems)
-                .collect(Collectors.toList());
-        } else {
-            donutData = Arrays.asList(0, 0, 0); // For A, B, C sections
-        }
-        chartData.put("donutData", donutData);
-        
-        // 2. Bar Chart Data - Aptitude Performance by Class
-        Map<String, Double> classAptitudeScores = batchStudents.stream()
-            .collect(Collectors.groupingBy(
-                Students::getClassName,
-                Collectors.averagingDouble(student -> 
-                    student.getStudentScores().stream()
-                        .mapToDouble(StudentScores::getScore)
-                        .average()
-                        .orElse(0.0)
-                )
-            ));
-        
-        List<Double> barData = new ArrayList<>(classAptitudeScores.values());
-        chartData.put("barData", barData);
-        
-        // 3. Line Chart Data - Weekly Performance Trends
-        Map<String, List<Double>> lineData = new HashMap<>();
-        
-        // LeetCode weekly averages
-        List<Double> leetcodeScores = new ArrayList<>();
-        for (int week = 1; week <= 6; week++) {
-            double weekAvg = batchStudents.stream()
-                .mapToDouble(Students::getLeetcodeScore)
-                .average()
-                .orElse(0.0);
-            leetcodeScores.add(weekAvg);
-        }
-        
-        // Aptitude weekly averages
-        List<Double> aptitudeScores = new ArrayList<>();
-        for (int week = 1; week <= 6; week++) {
-            double weekAvg = batchStudents.stream()
-                .flatMap(s -> s.getStudentScores().stream())
-                .mapToDouble(StudentScores::getScore)
-                .average()
-                .orElse(0.0);
-            aptitudeScores.add(weekAvg);
-        }
-        
-        // Overall weekly averages
-        List<Double> overallScores = new ArrayList<>();
-        for (int i = 0; i < 6; i++) {
-            double overall = (leetcodeScores.get(i) + aptitudeScores.get(i)) / 2.0;
-            overallScores.add(overall);
-        }
-        
-        lineData.put("leetcode", leetcodeScores);
-        lineData.put("aptitude", aptitudeScores);
-        lineData.put("overall", overallScores);
-        
-        chartData.put("lineData", lineData);
-        
-        return chartData;
-    }
-
-
-
-
-      /* function to get the student by the queryin using the params */
+      /* function to get the student by the querying using the params */
 
 
         public List<Students> searchStudents(String query) {
@@ -299,49 +206,8 @@ public class StudentsService {
 
         }
 
-     /* Method to get the toppers for the home page */
-
-        public ArrayList<Map<String, Object>> getToppers(int batch) {
-            ArrayList<Map<String, Object>> toppers = new ArrayList<>();
-            
-            /*Get the topper of the leetcode */
-
-            Students leetcodeTopper = studentsDB.findTopLeetcodeStudentByBatch(batch);
-            if (leetcodeTopper != null) {
-                Map<String, Object> leetcodeTopperMap = new HashMap<>();
-                leetcodeTopperMap.put("name", leetcodeTopper.getName());
-                leetcodeTopperMap.put("title", "LeetCode Topper");
-                leetcodeTopperMap.put("univNo", leetcodeTopper.getUniversityNo());
-                leetcodeTopperMap.put("score", leetcodeTopper.getLeetcodeScore());
-                leetcodeTopperMap.put("greets", "Congratulations " + leetcodeTopper.getName());
-                toppers.add(leetcodeTopperMap);
-            }
-            
-            /* Get Aptitude topper */
-
-            Students aptitudeTopper = studentsDB.findTopAptitudeStudentByBatch(batch);
-            if (aptitudeTopper != null) {
-                Map<String, Object> aptitudeTopperMap = new HashMap<>();
-                aptitudeTopperMap.put("name", aptitudeTopper.getName());
-                aptitudeTopperMap.put("title", "Aptitude Topper");
-                aptitudeTopperMap.put("univNo", aptitudeTopper.getUniversityNo());
-                aptitudeTopperMap.put("score", studentsDB.getAverageAptitudeScore(aptitudeTopper.getUniversityNo()));
-                aptitudeTopperMap.put("greets", "Congratulations " + aptitudeTopper.getName());
-                toppers.add(aptitudeTopperMap);
-            }
-            
-            return toppers;
-        }
-
-        
-
-    public List<Map<String, Object>> getAllAptitudeScores(int batch) {
-        return studentsDB.findAllAptitudeScoresByBatch(batch);
-    }
-
-    public List<Map<String, Object>> getAllLeetcodeScores(int batch) {
-        return studentsDB.findAllLeetcodeScoresByBatch(batch);
-    }
+ 
+    /* function to delete the students record */
 
     @Transactional
     public void deleteStudent(Long universityNo) {
@@ -354,25 +220,58 @@ public class StudentsService {
     /* Method to update the student score */
     
     @Transactional
-    public void updateStudentScore(Long universityNo, String testName, Double newScore) {
+    public Map<String, Object> updateStudentScore(Long universityNo, String testName, Double score) {
+        // Validate inputs
+        if (testName == null || testName.trim().isEmpty()) {
+            return Map.of("error", "Test name cannot be empty");
+        }
+        if (score == null) {
+            return Map.of("error", "Score cannot be null");
+        }
+
+        // Check if student exists
         Students student = studentsDB.findById(universityNo)
             .orElseThrow(() -> new StudentNotFoundException(universityNo));
-        
-        // Find if test score already exists
-        Optional<StudentScores> existingScore = student.getStudentScores().stream()
-            .filter(score -> score.getTestName().equals(testName))
-            .findFirst();
-        
-        if (existingScore.isPresent()) {
-            // Update existing score
-            existingScore.get().setScore(newScore);
-        } else {
-            // Add new score
-            StudentScores newStudentScore = new StudentScores(student, testName, newScore);
-            student.getStudentScores().add(newStudentScore);
+
+        try {
+            // Try to update existing score
+            int updatedRows = studentsDB.updateStudentTestScore(universityNo, testName, score);
+            
+            if (updatedRows == 0) {
+                // If no existing score found, create new one
+                student.addTestScore(testName, score);
+                studentsDB.save(student);
+            }
+
+            return Map.of(
+                "message", "Score updated successfully",
+                "universityNo", universityNo,
+                "testName", testName,
+                "score", score
+            );
+
+        } catch (Exception e) {
+            return Map.of("error", "Failed to update score: " + e.getMessage());
         }
-        
-        studentsDB.save(student);
+    }
+
+    /* Method to get the student scores */
+
+    public Map<String, Object> getStudentScores(Long universityNo) {
+        Students student = studentsDB.findById(universityNo)
+            .orElseThrow(() -> new StudentNotFoundException(universityNo));
+            
+        Map<String, Object> result = new HashMap<>();
+        result.put("universityNo", student.getUniversityNo());
+        result.put("name", student.getName());
+        result.put("scores", student.getStudentScores().stream()
+            .map(score -> Map.of(
+                "testName", score.getTestName(),
+                "score", score.getScore()
+            ))
+            .collect(Collectors.toList()));
+            
+        return result;
     }
 }
