+++ src/main/resources/application.properties
@@ -1,5 +1,18 @@
+# Application Name
 spring.application.name=automation
-spring.datasource.url=jdbc:h2:mem:test
+
+# MySQL Database Connection
+spring.datasource.url=jdbc:mysql://localhost:3306/spring
+spring.datasource.username=root
+spring.datasource.password=Gopi@4182
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+# Hibernate (JPA) Settings
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
-spring.datasource.username=Gopinath
-spring.datasource.password=123
+
+# Additional Debug Properties
+spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
